Indentation and Name
Formated the indentation, and name of file updated from longest-common-subsequence to lcs.
